adding rating/trajet scenario

+++ driver-service/src/main/java/esprit/driver/service/DriverService.java
@@ -28,7 +28,7 @@ public class DriverService {
 
 
 
-    public Trajet getrajetbyid(int id){
+    public Trajet getrajetbyid(Long id){
         return trajetServiceClient.getTrajetById(id);
     }
